Finish task11 and task12

+++ approximate.cpp
@@ -8,7 +8,7 @@ bool approximatelyEqual(double a, double b, double epsilon){
 }
 
 bool approximatelyEqualAbsRel(double a, double b, double absEpsilon, double relEpsilon) {
-	float diff = fabs(a - b);
+	double diff = fabs(a - b);
 	if (diff <= absEpsilon)
 		return true;
 	return diff <= ((fabs(a) < fabs(b) ? fabs(b) : fabs(a)) * relEpsilon);
@@ -18,18 +18,19 @@ bool approximatelyEqualAbsRel(double a, double b, double absEpsilon, double relE
 void doApproximate(int countArgs, char* args[]) {
 	if(countArgs == 5){
 		if(validArgsApproximate(countArgs, args)){
-			double num1 = std::stof(args[2]);
-			double num2 = std::stof(args[3]);
-			double epsilon= std::stof(args[4]);
+			double num1 = std::stod(args[2]);
+			double num2 = std::stod(args[3]);
+			double epsilon= std::stod(args[4]);
 			cout << "Numbers: " << num1 << " and " <<num2 << (approximatelyEqual(num1, num2, epsilon) ? " equals" : " no equals") << endl;
 		}
 	}
 	else if(countArgs == 6){
 		if(validArgsApproximate(countArgs, args)){
-			double num1 = std::stof(args[2]);
-			double num2 = std::stof(args[3]);
-			double absEpsilon = std::stof(args[4]);
-			double relEpsilon= std::stof(args[5]);
+			double num1 = std::stod(args[2]);
+			double num2 = std::stod(args[3]);
+			double absEpsilon = std::stod(args[4]);
+			double relEpsilon = std::stod(args[5]);
+			
 			cout << "Numbers: " << num1 << " and " <<num2 << (approximatelyEqualAbsRel(num1, num2, absEpsilon, relEpsilon) ? " equals" : " no equals") << endl;
 		}
 	}else cerr<<"Incorrect arguments"<<endl;
